// src/app/records/enums/add-emotion-step.enum.ts
export enum AddEmotionStep {
  SELECT_CORE = 'select core emotion',
  SELECT_SPECIFIC = 'select specific emotion',
  SELECT_FINE_GRAINED = 'select fine grained emotion',
  RATE = 'rate emotion',
  ADD_MORE_EMOTION = 'add more emotion',
  DONE_ADDING_EMOTION = 'done adding emotion',
  EXIT = 'exit'
}